+++ src/main/java/com/bytegriffin/datatunnel/read/KafkaConsumeReader.java
@@ -7,9 +7,6 @@ import java.util.Properties;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.apache.kafka.clients.consumer.ConsumerRecords;
 import org.apache.kafka.clients.consumer.KafkaConsumer;
-import org.apache.kafka.clients.producer.KafkaProducer;
-import org.apache.kafka.clients.producer.Producer;
-import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -79,26 +76,5 @@ public class KafkaConsumeReader implements Readable {
         }
         return list;
 	}
-
-	public static void main(String[] args) {
-		Properties props = new Properties();
-		props.put("bootstrap.servers", "localhost:9092");
-		props.put("acks", "all");
-		props.put("retries", 0);
-		props.put("batch.size", 16384);
-		props.put("linger.ms", 1);
-		props.put("buffer.memory", 33554432);
-		props.put("key.serializer", "org.apache.kafka.common.serialization.StringSerializer");
-		props.put("value.serializer", "org.apache.kafka.common.serialization.StringSerializer");
-
-		Producer<String, String> producer = new KafkaProducer<>(props);
-		for (int i = 0; i < 100; i++){
-			producer.send(new ProducerRecord<>("topic1", Integer.toString(i), Integer.toString(i)));
-			System.out.println("send成功");
-		}
-		
-		producer.close();
-	}
-
 
 }
