--- src/test/java/com/bytegriffin/datatunnel/Test.java
@@ -1,8 +1,0 @@
-package com.bytegriffin.datatunnel;
-
-public class Test {
-
-	public static void main(String[] args){
-		System.out.println("test");
-	}
-}
